requirements.txt
GitPython==3.1.14
gtrending==0.3.1
pip-check-reqs==2.2.0
rich==9.13.0
